+++ release/m/masaram_gondi/source/help/masaram_gondi.php
@@ -1,0 +1,27 @@
+<?php 
+  $pagename = 'Masaram Gondi Phonetic Keyboard Help';
+  $pagetitle = $pagename;
+  // Header we will tidy up later  
+  require_once('header.php');
+?>
+
+
+<p>
+This keyboard is designed for typing the Masaram Gondi language in the Gondi script. 
+A Gondi font is required for displaying the characters. This keyboard package 
+will install the Noto Sans MasaramGondi-Regular font.
+</p>
+<p>You can phonetically type Gondi 
+    using the modified ITRANS transliteration scheme on a regular QWERTY keyboard
+    and get Masaram Gondi output. </p>
+
+<h1>Keyboard Layout</h1>
+
+<h2>Desktop Keyboard Layout</h2>
+<p>The default and shifted layouts cover Devanagari digits, consonants, vowels and other commonly used characters.</p>
+<center>
+<h3>Default (unshifted)</h3>
+<p><img src="gondi_keyboard_without_shift.jpg" alt="Default (unshifted) state"/></p>
+<h3>Shift</h3>
+<p><img src="gondi_keyboard_with_shift.jpg" alt="Shift state"/></p>
+</center>